Split APIConfigurationImages into seperate file
