+++ TMDbLib/Objects/Configuration/APIConfiguration.cs
@@ -11,28 +11,4 @@ namespace TMDbLib.Objects.Configuration
         [JsonProperty("change_keys")]
         public List<string> ChangeKeys { get; set; }
     }
-
-    public class APIConfigurationImages
-    {
-        [JsonProperty("base_url")]
-        public string BaseUrl { get; set; }
-
-        [JsonProperty("secure_base_url")]
-        public string SecureBaseUrl { get; set; }
-
-        [JsonProperty("backdrop_sizes")]
-        public List<string> BackdropSizes { get; set; }
-
-        [JsonProperty("logo_sizes")]
-        public List<string> LogoSizes { get; set; }
-
-        [JsonProperty("poster_sizes")]
-        public List<string> PosterSizes { get; set; }
-
-        [JsonProperty("profile_sizes")]
-        public List<string> ProfileSizes { get; set; }
-
-        [JsonProperty("still_sizes")]
-        public List<string> StillSizes { get; set; }
-    }
 }
